requirements.txt
Django==1.8
django-crispy-forms==1.6.0
django-filter==1.1.0
djangorestframework==3.5.3
Markdown==2.6.7
mysql-connector==2.2.9
mysqlclient==1.4.6
pytz==2019.3
